(core:subscriptions) print errors with context
Subscriptions execute outside of the module context. Whilst the proc is
scoped inside the module the error is handled by the global logger. So
errors occurring inside a subscription would be logged however would
not appear in module debug messages.
Now the errors are handled in the context of the subscriber. This fixes
#6