(core:proxy) make `.all(:Driver)` return an enumerable
Adds a send method which performs the same role as method_missing.
Send, as defined in the drivers, is protected from cross driver use so
it makes sense to use this method name on proxy objects.

Making a driver collection a proper enumerable pollutes the namespace a
little bit however send can be used if there is a name clash